control-plane/requirements.txt
# TazoSploit SaaS v2 - Control Plane Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0


# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
prometheus-client>=0.19.0

# Task queue
celery>=5.3.0

# Scheduling
croniter>=2.0.0
pytz>=2023.3

# Docker (for terminal sessions)
docker>=7.0.0

# MinIO (for evidence storage)
minio>=7.2.0

# PDF report generation
reportlab>=4.0.0

# HTTP client (for API key testing)
httpx>=0.27.0
